Handle deletion of ghost/placeholder football matches
for some competitions PA issue placeholder matches for KO stage then
delete them and replace with real matches when the teams are
known. They'll be deleting these matches and re-issuing them with real
teams when they become known. To prevent duplicate matches existing in
memory we need to remove the placeholders from the agent when they are
replaced.